requirements.txt
playwright
requests
beautifulsoup4
transformers
sentence-transformers
scikit-learn
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pytest
httpx
google-generativeai
